+++ src/configs/index.d.ts
@@ -1,3 +1,5 @@
 import type { Config } from 'stylelint';
 
+export const accessibility: Config;
 export const recommended: Config;
+export const strict: Config;