+++ tsdown.config.ts
@@ -9,6 +9,6 @@ export default defineConfig({
     index: 'src/index.ts',
   },
   exports: true,
-  external: ['stylelint'],
+  external: ['csstype', 'stylelint'],
   format: ['cjs', 'es'],
 });